requirements.txt
boto3==1.28.28
botocore==1.31.28
load-dotenv==0.1.0
python-dotenv==1.0.0
s3transfer==0.6.2